requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
docopt==0.6.2
fastapi==0.127.0
filelock==3.20.1
flatbuffers==25.12.19
fsspec==2025.12.0
h11==0.16.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
loguru==0.7.3
mpmath==1.3.0
numpy==2.4.0
onnxruntime==1.23.2
packaging==25.0
pillow==12.0.0
pipreqs==0.4.13
protobuf==6.33.2
psutil==7.2.0
pydantic==2.12.5
pydantic_core==2.41.5
pyreadline3==3.5.4
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
win32_setctime==1.2.0
yarg==0.1.10
